add exception handling for department deletion when employees are present

+++ src/main/java/dev/jonkursani/restapigr2/services/impls/DepartmentServiceImpl.java
@@ -4,6 +4,7 @@ import dev.jonkursani.restapigr2.dtos.department.CreateDepartmentRequest;
 import dev.jonkursani.restapigr2.dtos.department.DepartmentDto;
 import dev.jonkursani.restapigr2.dtos.department.DepartmentWithEmployeeCount;
 import dev.jonkursani.restapigr2.dtos.department.UpdateDepartmentRequest;
+import dev.jonkursani.restapigr2.exceptions.department.DepartmentHasEmployeesException;
 import dev.jonkursani.restapigr2.exceptions.department.DepartmentNotFoundException;
 import dev.jonkursani.restapigr2.mappers.DepartmentMapper;
 import dev.jonkursani.restapigr2.repositories.DepartmentRepository;
@@ -54,9 +55,14 @@ public class DepartmentServiceImpl implements DepartmentService {
 
     @Override
     public void delete(Integer id) {
-//        var department = repository.findById(id)
-//                .orElseThrow(() -> new DepartmentNotFoundException(id));
-        findById(id);
+        var department = repository.findById(id)
+                .orElseThrow(() -> new DepartmentNotFoundException(id));
+//        findById(id);
+
+        if (!department.getEmployees().isEmpty()) {
+            throw new DepartmentHasEmployeesException(id);
+        }
+
         repository.deleteById(id);
     }
